Remove obsolete file.
This file is no longer used to install the dependencies. The depen-
dencies are installed by `install.sh` in the **releaser** application,

The "installation" is simply copying the repository into the
`~/.babushka/deps/...` directory which should be sufficient.

The other part of this file (resetting the Babushka installation to a
known good Git ref) has also been removed as the Babushka framework
will always be installed from the included tarball.